Added: Missing 'UInt32' types in attribute definitions

+++ src/Abstractions/NexusMods.Abstractions.NexusWebApi/Types/V2/FileId.cs
@@ -21,7 +21,7 @@ public readonly partial struct FileId : IAugmentWith<DefaultValueAugment>, IAugm
 /// File ID attribute, for NexusMods API file IDs.
 /// </summary>
 public class FileIdAttribute(string ns, string name) : 
-    ScalarAttribute<FileId, uint>(ValueTags.UInt64, ns, name)
+    ScalarAttribute<FileId, uint>(ValueTags.UInt32, ns, name)
 {
     /// <inheritdoc />
     protected override uint ToLowLevel(FileId value) => value.Value;
